// frontend/src/app/core/generated/circabc/model/columnOptions.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ColumnOptions { 
    name?: boolean;
    title?: boolean;
    version?: boolean;
    modification?: boolean;
    creation?: boolean;
    size?: boolean;
    expiration?: boolean;
    status?: boolean;
    description?: boolean;
    author?: boolean;
    securityRanking?: boolean;
}