// frontend/src/app/core/generated/circabc/model/pagedNodes.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Node } from './node';


export interface PagedNodes { 
    data: Array<Node>;
    /**
     * return the total amount of nodes, without paging so the UI can compute the number of pages 
     */
    total: number;
}