// frontend/src/app/core/generated/circabc/model/preferenceConfiguration.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SearchPreferences } from './searchPreferences';
import { LibraryPreferences } from './libraryPreferences';


export interface PreferenceConfiguration { 
    library: LibraryPreferences;
    search: Array<SearchPreferences>;
}